Refactor to use util package
