+++ smrt-server-lims/src/test/scala/com/pacbio/secondary/lims/RouteImportAndResloveSpec.scala
@@ -5,8 +5,8 @@ import com.pacbio.secondary.lims.services.{ImportLimsYml, ResolveDataSet}
 import org.specs2.mutable.Specification
 import spray.http._
 import spray.testkit.Specs2RouteTest
-
 import com.pacbio.secondary.lims.JsonProtocol._
+import com.pacbio.secondary.lims.util.StressUtil
 import spray.json.DefaultJsonProtocol._
 import spray.json._
